chore: update Gemfile.lock dependencies
This commit updates the dependencies in Gemfile.lock to their latest versions. This includes updates to aws-partitions, oj, rubocop, rubocop-ast, rubocop-rails, rubocop-rails-omakase and uri. These updates are necessary to ensure that the application is using the latest versions of these libraries, which may include bug fixes, security patches, and new features.